+++ src/Components/Base/BaseText.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable react-native/no-inline-styles */
 import React, { useMemo } from "react"
-import { FlexAlignType, Text, TextProps } from "react-native"
-import { useTheme } from "~Hooks"
+import { FlexAlignType, Text, TextProps, TextStyle } from "react-native"
 import { typography } from "~Constants/Theme"
-import { BaseView } from "./BaseView"
+import { useTheme } from "~Hooks"
+import { BaseView, BaseViewProps } from "./BaseView"
 
 const { defaults: defaultTypography, ...otherTypography } = typography
 
@@ -36,10 +36,12 @@ export type BaseTextProps = {
     borderRadius?: number
     alignContainer?: FlexAlignType
     justifyContainer?: "flex-start" | "flex-end" | "center" | "space-between" | "space-around" | "space-evenly"
-} & TextProps
+    textTransform?: TextStyle["textTransform"]
+} & TextProps &
+    Pick<BaseViewProps, "flex" | "flexGrow" | "flexShrink">
 
 export const BaseText = (props: BaseTextProps) => {
-    const { style, typographyFont, fontSize, fontWeight, fontFamily, lineHeight, ...otherProps } = props
+    const { style, typographyFont, fontSize, fontWeight, fontFamily, lineHeight, textTransform, ...otherProps } = props
     const theme = useTheme()
 
     const computedFontSize = useMemo(
@@ -88,7 +90,10 @@ export const BaseText = (props: BaseTextProps) => {
             w={props.w}
             borderRadius={props.borderRadius}
             bg={props.bg}
-            h={props.h}>
+            h={props.h}
+            flex={props.flex}
+            flexGrow={props.flexGrow}
+            flexShrink={props.flexShrink}>
             <Text
                 style={[
                     {
@@ -99,6 +104,7 @@ export const BaseText = (props: BaseTextProps) => {
                         textAlign: props.align,
                         fontStyle: props.italic ? "italic" : "normal",
                         textDecorationLine: props.underline ? "underline" : "none",
+                        textTransform,
                         lineHeight:
                             lineHeight ?? (typographyFont ? defaultTypography[typographyFont].lineHeight : undefined),
                     },