+++ src/Components/Base/BaseToast/components/ToastAddressesContent.tsx
@@ -1,12 +1,12 @@
 import React, { useMemo } from "react"
-import { ToastAddress } from "../BaseToast"
-import { ToastStyles } from "../util"
+import { StyleSheet } from "react-native"
 import { BaseIcon, BaseText, BaseView } from "~Components"
 import { useThemedStyles, useVns } from "~Hooks"
 import { selectAccountByAddress, useAppSelector } from "~Storage/Redux"
-import { formatAlias } from "~Utils/FormattingUtils/FormattingUtils"
 import { humanAddress } from "~Utils/AddressUtils/AddressUtils"
-import { StyleSheet } from "react-native"
+import { formatAlias } from "~Utils/FormattingUtils/FormattingUtils"
+import { ToastAddress } from "../BaseToast"
+import { ToastStyles } from "../util"
 
 type Props = {
     addresses: ToastAddress
@@ -24,14 +24,14 @@ export const ToastAddressesContent: React.FC<Props> = ({ addresses, styles }) =>
     const senderTitle = useMemo(() => {
         if (senderAccount && !senderName) return formatAlias(senderAccount.alias, 18, 6, 8)
         if (senderName) return formatAlias(senderName, 18, 6, 8)
-        return humanAddress(addresses.sender, 4, 6)
+        return humanAddress(addresses.sender)
     }, [addresses.sender, senderAccount, senderName])
 
     const recipientTitle = useMemo(() => {
         if (!addresses.recipient) return ""
         if (recipientAccount && !recipientName) return formatAlias(recipientAccount.alias, 18, 6, 8)
         if (recipientName) return formatAlias(recipientName, 18, 6, 8)
-        return humanAddress(addresses.recipient, 4, 6)
+        return humanAddress(addresses.recipient)
     }, [addresses.recipient, recipientAccount, recipientName])
 
     return (