+++ src/Screens/Flows/App/DiscoverScreen/DiscoverScreen.tsx
@@ -1,6 +1,6 @@
 import { useNavigation, useScrollToTop } from "@react-navigation/native"
 import React, { useCallback, useEffect, useMemo, useRef, useState } from "react"
-import { Keyboard, Linking, StyleSheet } from "react-native"
+import { Keyboard, StyleSheet } from "react-native"
 import Animated, { useAnimatedRef, useScrollViewOffset } from "react-native-reanimated"
 import {
     AnimatedFloatingButton,
@@ -25,17 +25,9 @@ import {
 } from "~Storage/Redux"
 import { URIUtils } from "~Utils"
 import { useI18nContext } from "~i18n"
-import {
-    AnimatedTitle,
-    Ecosystem,
-    Favourites,
-    MakeYourOwnDApp,
-    VeBetterDAODApps,
-    VeBetterDAOMainCard,
-} from "./Components"
+import { AnimatedTitle, Ecosystem, Favourites, VeBetterDAODApps } from "./Components"
 import { groupFavoritesByBaseUrl } from "./utils"
-
-const DAO_URL = "https://governance.vebetterdao.org/"
+import { VeBetterDAOCarousel } from "./Components/VeBetterDAOCarousel"
 
 export const DiscoverScreen: React.FC = () => {
     const { LL } = useI18nContext()
@@ -109,13 +101,6 @@ export const DiscoverScreen: React.FC = () => {
         setFilteredSearch("")
     }, [dapps, filteredSearch, navigateToBrowser])
 
-    const onMakeYourOwnDAppPress = useCallback(async () => {
-        const url = process.env.REACT_APP_CREATE_YOUR_VECHAIN_DAPP_URL
-        if (url && (await Linking.canOpenURL(url))) {
-            Linking.openURL(url)
-        }
-    }, [])
-
     const onNavigateToBrowserHistory = useCallback(() => {
         nav.navigate(Routes.DISCOVER_BROWSER_HISTORY)
     }, [nav])
@@ -169,11 +154,9 @@ export const DiscoverScreen: React.FC = () => {
                         showsHorizontalScrollIndicator={false}
                         showsVerticalScrollIndicator={false}>
                         <BaseSpacer height={24} />
-                        <VeBetterDAOMainCard href={DAO_URL} onDAppPress={onDAppPress} />
+                        <VeBetterDAOCarousel />
                         <BaseSpacer height={36} />
                         <VeBetterDAODApps onDAppPress={onDAppPress} />
-                        <BaseSpacer height={42} />
-                        <MakeYourOwnDApp label={LL.DISCOVER_CREATE_YOUR_DAPP()} onPress={onMakeYourOwnDAppPress} />
                         <BaseSpacer height={18} />
                         {showFavorites && (
                             <Favourites